requirements.txt
pyyaml>=6.0
aiohttp>=3.9
dnspython>=2.5
requests>=2.31
